requirements.txt
Django==4.0.10
djangorestframework==3.13.0
python-dotenv==1.0.0
drf-yasg==1.21.7
psycopg2
